pyproject.toml
[project]
name = "pipepair"
version = "0.1.0"
description = "Crash-resistant pipeline executor with auto-repair"
authors = [{ name="fzrrahmani", email="fzrrahmani@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.7"
dependencies = ["pyyaml"]

[project.scripts]
pipepair = "pipepair.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"